requirements.txt
discord.py==2.4.0
python-dotenv==1.1.0
Requests==2.32.3
